+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
-
+### Added
+- Method to calculate revolutions from pulse count and steps per revolution
 
 ##  [0.9.10]
 ### Added